SL-613, refactor code based on the reviews
